#2 Fikset problem med at selve applikasjonen ikke blir blokkert. ProtectedRoutes er kommentert ut. Flyttet createBrowserRouter til main.tsx.
